move entities into 'models'
